đánh giá sản phẩm

+++ be/src/main/java/com/project/pharmacy/repository/OrderRepository.java
@@ -2,9 +2,12 @@ package com.project.pharmacy.repository;
 
 import com.project.pharmacy.entity.Orders;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
 import java.util.*;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Orders, Long> {
+public interface OrderRepository extends JpaRepository<Orders, Integer> {
 }
